backend/requirements.txt
flask
requests
flask-cors
facenet-pytorch==2.5.2
easydict==1.9
numpy>=1.19.3
tqdm
# opencv_python
Pillow==7.1.2
tensorboardX==2.0
pandas
opencv-python==4.5.1.48
matplotlib
seaborn
scikit-learn
torch
torchvision  
git+https://github.com/albumentations-team/albumentations